refactor(core): split SocketUtil into focused modules
Split SocketUtil.c/h into three focused modules for improved
maintainability and reduced compilation dependencies:

- SocketError.c/.h - Error handling and errno categorization
- SocketLog.c/.h - Logging subsystem with callbacks and levels
- SocketEvent.c/.h - Event dispatching system

SocketUtil.h remains as an umbrella header that includes the
split modules for backward compatibility. All inline utilities,
macros, and legacy metrics remain in SocketUtil.h/c.

Changes:
- Created include/core/SocketError.h with error handling APIs
- Created src/core/SocketError.c with errno mapping and categorization
- Created include/core/SocketLog.h with logging APIs
- Created src/core/SocketLog.c with logging implementation
- Created include/core/SocketEvent.h with event system APIs
- Created src/core/SocketEvent.c with event dispatching
- Updated SocketUtil.h to include split modules
- Updated SocketUtil.c to keep time utilities and legacy metrics
- Updated CMakeLists.txt to build new source files

Benefits:
- Reduced compilation dependencies (can include only what's needed)
- Focused testing for each subsystem
- Independent evolution of modules
- Clearer ownership and responsibility
- Backward compatible via umbrella header

Testing:
- All 111 tests pass with sanitizers (ASan/UBSan)
- No breaking changes to existing code

Closes #502
