+++ include/core/SocketUtil.h
@@ -47,391 +47,157 @@
 #include "core/SocketConfig.h"
 
 /* ============================================================================
- * LOGGING SUBSYSTEM
+ * CONSOLIDATED INCLUDES (Split Modules)
+ * ============================================================================
+ *
+ * These subsystems have been split into focused modules for improved
+ * maintainability and reduced compilation dependencies. SocketUtil.h
+ * remains as an umbrella header for backward compatibility.
+ */
+
+#include "core/SocketError.h"  /* Error handling and categorization */
+#include "core/SocketLog.h"    /* Logging subsystem */
+#include "core/SocketEvent.h"  /* Event dispatching */
+
+/* ============================================================================
+ * ERROR HANDLING MACROS (Combine Error + Log)
  * ============================================================================
  */
 
 /**
- * @brief Log severity levels.
- * @ingroup foundation
+ * @brief SOCKET_ERROR_FMT - Format error message with errno information
+ *
+ * Includes truncation protection for long messages.
  */
-typedef enum SocketLogLevel
-{
-  SOCKET_LOG_TRACE = 0,
-  SOCKET_LOG_DEBUG,
-  SOCKET_LOG_INFO,
-  SOCKET_LOG_WARN,
-  SOCKET_LOG_ERROR,
-  SOCKET_LOG_FATAL
-} SocketLogLevel;
+#define SOCKET_ERROR_FMT(fmt, ...)                                            \
+  do                                                                          \
+    {                                                                         \
+      socket_last_errno = errno;                                              \
+      char tmp_buf[SOCKET_ERROR_BUFSIZE];                                     \
+      int _socket_error_ret = snprintf (                                      \
+          tmp_buf, sizeof (tmp_buf), fmt " (errno: %d - %s)", ##__VA_ARGS__,  \
+          socket_last_errno, Socket_safe_strerror (socket_last_errno));       \
+      memcpy (socket_error_buf, tmp_buf, SOCKET_ERROR_BUFSIZE);               \
+      socket_error_buf[SOCKET_ERROR_BUFSIZE - 1] = '\0';                      \
+      SOCKET_ERROR_APPLY_TRUNCATION (_socket_error_ret);                      \
+      (void)_socket_error_ret;                                                \
+      SocketLog_emit (SOCKET_LOG_ERROR, SOCKET_LOG_COMPONENT,                 \
+                      socket_error_buf);                                      \
+    }                                                                         \
+  while (0)
 
 /**
- * @brief Custom logging callback function type.
- * @ingroup foundation
- * @param userdata User-provided context.
- * @param level Log severity level.
- * @param component Module/component name.
- * @param message Log message.
- * @see SocketLog_setcallback() for registration.
+ * @brief SOCKET_ERROR_MSG - Format error message without errno
+ *
+ * Includes truncation protection for long messages.
  */
-typedef void (*SocketLogCallback) (void *userdata, SocketLogLevel level,
-                                   const char *component, const char *message);
-
-/**
- * @brief Register a custom callback for all library log emissions.
- * @ingroup foundation
- *
- * @param callback Callback function or NULL for default logger.
- * @param userdata Opaque user data passed to callback.
- *
- * Overrides default stdout/stderr logging. Callback invoked synchronously
- * from emitting thread after level filtering. Keep callbacks non-blocking.
- *
- * @threadsafe Yes
- */
-void SocketLog_setcallback (SocketLogCallback callback, void *userdata);
-
-/**
- * @brief Retrieve the currently registered logging callback and userdata.
- * @ingroup foundation
- *
- * @param userdata Output for userdata (may be NULL).
- * @return Current SocketLogCallback, or internal default if none registered.
- *
- * @threadsafe Yes
- */
-SocketLogCallback SocketLog_getcallback (void **userdata);
-
-/**
- * @brief Get human-readable string for a log level.
- * @ingroup foundation
- *
- * @param level Log level enum value.
- * @return Static string ("TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL").
- *
- * @threadsafe Yes
- */
-const char *SocketLog_levelname (SocketLogLevel level);
-
-/**
- * @brief Emit a plain log message.
- * @ingroup foundation
- *
- * @param level Log severity level.
- * @param component Module name (may be NULL).
- * @param message Log message (may be NULL).
- *
- * @threadsafe Yes
- */
-void SocketLog_emit (SocketLogLevel level, const char *component,
-                     const char *message);
-
-/**
- * @brief Emit a formatted log message (printf-style).
- * @ingroup foundation
- *
- * @param level Log level.
- * @param component Component name.
- * @param fmt Printf-style format string.
- * @param ... Format arguments.
- *
- * @warning fmt must be a compile-time literal to prevent format string attacks.
- * @threadsafe Yes
- */
-void SocketLog_emitf (SocketLogLevel level, const char *component,
-                      const char *fmt, ...)
-    __attribute__ ((format (printf, 3, 4)));
-
-/**
- * @brief Emit formatted log message using va_list.
- * @ingroup foundation
- *
- * @param level Log level.
- * @param component Component name.
- * @param fmt Printf-style format string.
- * @param args Format arguments as va_list.
- *
- * @threadsafe Yes
- */
-void SocketLog_emitfv (SocketLogLevel level, const char *component,
-                       const char *fmt, va_list args)
-    __attribute__ ((format (printf, 3, 0)));
-
-/**
- * @brief Configure global minimum log level threshold.
- * @ingroup foundation
- *
- * @param min_level Minimum level (SOCKET_LOG_TRACE = most verbose).
- *
- * Logs below this level are suppressed. Default: SOCKET_LOG_INFO.
- *
- * @threadsafe Yes
- */
-extern void SocketLog_setlevel (SocketLogLevel min_level);
-
-/**
- * @brief Get the current global minimum log level threshold.
- * @ingroup foundation
- *
- * @return Current SocketLogLevel threshold.
- *
- * @threadsafe Yes
- */
-extern SocketLogLevel SocketLog_getlevel (void);
-
-/* Default log component - modules should override before including this header
- */
-#ifndef SOCKET_LOG_COMPONENT
-#define SOCKET_LOG_COMPONENT "Socket"
-#endif
-
-/* ----------------------------------------------------------------------------
- * Convenience Logging Macros
- * ----------------------------------------------------------------------------
- *
- * These macros provide ergonomic logging that automatically uses the
- * SOCKET_LOG_COMPONENT macro defined by each module. Each module should
- * define SOCKET_LOG_COMPONENT before including this header:
- *
- *   #undef SOCKET_LOG_COMPONENT
- *   #define SOCKET_LOG_COMPONENT "MyModule"
- *
- * Usage:
- *   SOCKET_LOG_DEBUG_MSG("Connection established fd=%d", fd);
- *   SOCKET_LOG_ERROR_MSG("Failed to bind: %s", strerror(errno));
- */
-
-/* Log at TRACE level (most verbose, detailed tracing) */
-#define SOCKET_LOG_TRACE_MSG(fmt, ...)                                        \
-  SocketLog_emitf (SOCKET_LOG_TRACE, SOCKET_LOG_COMPONENT, fmt, ##__VA_ARGS__)
-
-/* Log at DEBUG level (debugging information) */
-#define SOCKET_LOG_DEBUG_MSG(fmt, ...)                                        \
-  SocketLog_emitf (SOCKET_LOG_DEBUG, SOCKET_LOG_COMPONENT, fmt, ##__VA_ARGS__)
-
-/* Log at INFO level (normal operational messages) */
-#define SOCKET_LOG_INFO_MSG(fmt, ...)                                         \
-  SocketLog_emitf (SOCKET_LOG_INFO, SOCKET_LOG_COMPONENT, fmt, ##__VA_ARGS__)
-
-/* Log at WARN level (warning conditions) */
-#define SOCKET_LOG_WARN_MSG(fmt, ...)                                         \
-  SocketLog_emitf (SOCKET_LOG_WARN, SOCKET_LOG_COMPONENT, fmt, ##__VA_ARGS__)
-
-/* Log at ERROR level (error conditions) */
-#define SOCKET_LOG_ERROR_MSG(fmt, ...)                                        \
-  SocketLog_emitf (SOCKET_LOG_ERROR, SOCKET_LOG_COMPONENT, fmt, ##__VA_ARGS__)
-
-/* Log at FATAL level (critical errors, typically before abort) */
-#define SOCKET_LOG_FATAL_MSG(fmt, ...)                                        \
-  SocketLog_emitf (SOCKET_LOG_FATAL, SOCKET_LOG_COMPONENT, fmt, ##__VA_ARGS__)
-
-/* ----------------------------------------------------------------------------
- * Safe Logging Macros (for untrusted/user-controlled messages)
- * ----------------------------------------------------------------------------
- *
- * These macros use SocketLog_emit() with a fixed "%s" format to safely log
- * user-controlled strings without format string vulnerabilities.
- *
- * Use these when logging data that may come from untrusted sources.
- *
- * Usage:
- *   const char *user_input = get_user_input();
- *   SOCKET_LOG_INFO_SAFE(user_input);  // Safe - no format string attack
- */
-
-/* Log untrusted string at TRACE level */
-#define SOCKET_LOG_TRACE_SAFE(msg)                                            \
-  SocketLog_emit (SOCKET_LOG_TRACE, SOCKET_LOG_COMPONENT, (msg))
-
-/* Log untrusted string at DEBUG level */
-#define SOCKET_LOG_DEBUG_SAFE(msg)                                            \
-  SocketLog_emit (SOCKET_LOG_DEBUG, SOCKET_LOG_COMPONENT, (msg))
-
-/* Log untrusted string at INFO level */
-#define SOCKET_LOG_INFO_SAFE(msg)                                             \
-  SocketLog_emit (SOCKET_LOG_INFO, SOCKET_LOG_COMPONENT, (msg))
-
-/* Log untrusted string at WARN level */
-#define SOCKET_LOG_WARN_SAFE(msg)                                             \
-  SocketLog_emit (SOCKET_LOG_WARN, SOCKET_LOG_COMPONENT, (msg))
-
-/* Log untrusted string at ERROR level */
-#define SOCKET_LOG_ERROR_SAFE(msg)                                            \
-  SocketLog_emit (SOCKET_LOG_ERROR, SOCKET_LOG_COMPONENT, (msg))
-
-/* Log untrusted string at FATAL level */
-#define SOCKET_LOG_FATAL_SAFE(msg)                                            \
-  SocketLog_emit (SOCKET_LOG_FATAL, SOCKET_LOG_COMPONENT, (msg))
-
-/* ----------------------------------------------------------------------------
- * Thread-Local Logging Context
- * ----------------------------------------------------------------------------
- *
- * Provides correlation IDs for distributed tracing and request tracking.
- * Each thread can set its own context that will be available to custom
- * logging callbacks for inclusion in log output.
- *
- * Usage:
- *   SocketLogContext ctx = {0};
- *   strncpy(ctx.trace_id, "abc-123-def", sizeof(ctx.trace_id) - 1);
- *   ctx.connection_fd = client_fd;
- *   SocketLog_setcontext(&ctx);
- *
- *   // ... handle request - all logs will have context available ...
- *
- *   SocketLog_clearcontext();
- */
-
-/* UUID size: 36 chars (8-4-4-4-12) + NUL */
-#define SOCKET_LOG_ID_SIZE 37
-
-/**
- * @brief Thread-local logging context for distributed tracing.
- * @ingroup foundation
- */
-typedef struct SocketLogContext
-{
-  char trace_id[SOCKET_LOG_ID_SIZE]; /**< Distributed trace ID (e.g., UUID) */
-  char request_id[SOCKET_LOG_ID_SIZE]; /**< Request-specific ID */
-  int connection_fd; /**< Associated file descriptor (-1 if none) */
-} SocketLogContext;
-
-/**
- * @brief Set thread-local logging context.
- * @ingroup foundation
- *
- * @param ctx Context to copy (NULL clears context).
- *
- * @threadsafe Yes (thread-local storage)
- */
-extern void SocketLog_setcontext (const SocketLogContext *ctx);
-
-/**
- * @brief Get current thread's logging context.
- * @ingroup foundation
- *
- * @return Pointer to thread-local SocketLogContext or NULL if unset.
- *
- * @threadsafe Yes
- */
-extern const SocketLogContext *SocketLog_getcontext (void);
-
-/**
- * @brief Clear thread-local logging context.
- * @ingroup foundation
- *
- * @threadsafe Yes
- */
-extern void SocketLog_clearcontext (void);
-
-/* ----------------------------------------------------------------------------
- * Structured Logging
- * ----------------------------------------------------------------------------
- *
- * Provides key-value pair logging for machine-parseable output.
- * Custom callbacks can format these fields as JSON, logfmt, etc.
- *
- * Usage:
- *   SocketLogField fields[] = {
- *       {"fd", "42"},
- *       {"bytes", "1024"},
- *       {"peer", "192.168.1.1"}
- *   };
- *   SocketLog_emit_structured(SOCKET_LOG_INFO, "Socket",
- *                             "Connection established",
- *                             fields, 3);
- *
- * Or with the convenience macro:
- *   SocketLog_emit_structured(SOCKET_LOG_INFO, "Socket",
- *                             "Connection established",
- *                             SOCKET_LOG_FIELDS(
- *                                 {"fd", "42"},
- *                                 {"bytes", "1024"}
- *                             ));
- */
-
-/**
- * @brief Key-value pair for structured logging.
- * @ingroup foundation
- */
-typedef struct SocketLogField
-{
-  const char *key;   /**< Field name (e.g., "fd", "bytes", "peer") */
-  const char *value; /**< Field value as string */
-} SocketLogField;
-
-/**
- * @brief Callback for structured logging with key-value fields.
- * @ingroup foundation
- *
- * @param userdata User-provided context.
- * @param level Log severity level.
- * @param component Module name.
- * @param message Log message.
- * @param fields Array of key-value pairs (may be NULL).
- * @param field_count Number of fields.
- * @param context Thread logging context (may be NULL).
- */
-typedef void (*SocketLogStructuredCallback) (
-    void *userdata, SocketLogLevel level, const char *component,
-    const char *message, const SocketLogField *fields, size_t field_count,
-    const SocketLogContext *context);
-
-/**
- * @brief Register callback for handling structured log emissions with fields.
- * @ingroup foundation
-
- * @callback: Callback function or NULL to disable structured logging
- * @userdata: User data passed to callback
- *
- * @brief Thread-safe: Yes (mutex protected)
-
- *
- * When set, SocketLog_emit_structured() will invoke this callback
- * instead of the regular callback, providing access to structured fields.
- */
-extern void
-SocketLog_setstructuredcallback (SocketLogStructuredCallback callback,
-                                 void *userdata);
-
-/**
- * @brief Emit log message with attached structured key-value metadata fields.
- * @ingroup foundation
-
- * @level: Log level
- * @component: Component name
- * @message: Log message
- * @fields: Array of key-value pairs (may be NULL)
- * @field_count: Number of fields
- *
- * @threadsafe Yes
-
- *
- * Emits a log message with structured key-value pairs. If a structured
- * callback is set, it receives the fields directly. Otherwise, fields
- * are formatted as "key=value" pairs appended to the message.
- */
-extern void SocketLog_emit_structured (SocketLogLevel level,
-                                       const char *component,
-                                       const char *message,
-                                       const SocketLogField *fields,
-                                       size_t field_count);
-
-/**
- * @brief SOCKET_LOG_FIELDS - Convenience macro for creating field arrays
-
- *
- * Usage:
- *   SocketLog_emit_structured(level, component, message,
- *                             SOCKET_LOG_FIELDS({"key1", "val1"},
- *                                               {"key2", "val2"}));
- */
-#define SOCKET_LOG_FIELDS(...)                                                \
-  (SocketLogField[]){ __VA_ARGS__ },                                          \
-      (sizeof ((SocketLogField[]){ __VA_ARGS__ }) / sizeof (SocketLogField))
+#define SOCKET_ERROR_MSG(fmt, ...)                                            \
+  do                                                                          \
+    {                                                                         \
+      socket_last_errno = errno;                                              \
+      char tmp_buf[SOCKET_ERROR_BUFSIZE];                                     \
+      int _socket_error_ret                                                   \
+          = snprintf (tmp_buf, sizeof (tmp_buf), fmt, ##__VA_ARGS__);         \
+      memcpy (socket_error_buf, tmp_buf, SOCKET_ERROR_BUFSIZE);               \
+      socket_error_buf[SOCKET_ERROR_BUFSIZE - 1] = '\0';                      \
+      SOCKET_ERROR_APPLY_TRUNCATION (_socket_error_ret);                      \
+      (void)_socket_error_ret;                                                \
+      SocketLog_emit (SOCKET_LOG_ERROR, SOCKET_LOG_COMPONENT,                 \
+                      socket_error_buf);                                      \
+    }                                                                         \
+  while (0)
 
 /* ============================================================================
- * METRICS SUBSYSTEM
+ * Centralized Exception Infrastructure
+ * ============================================================================
+ */
+
+/**
+ * @brief SOCKET_DECLARE_MODULE_EXCEPTION - Declare thread-local exception
+ *
+ * @module_name: Module name (e.g., Socket, SocketBuf, SocketPoll)
+ */
+#define SOCKET_DECLARE_MODULE_EXCEPTION(module_name)                          \
+  static __thread Except_T module_name##_DetailedException
+
+/**
+ * @brief SOCKET_RAISE_MODULE_ERROR - Raise module-specific exception
+ *
+ * @module_name: Module name
+ * @exception: Exception to raise
+ * @brief Thread-safe: Creates thread-local copy with detailed reason
+ *
+ */
+#define SOCKET_RAISE_MODULE_ERROR(module_name, exception)                     \
+  do                                                                          \
+    {                                                                         \
+      module_name##_DetailedException = (exception);                          \
+      module_name##_DetailedException.reason = socket_error_buf;              \
+      RAISE (module_name##_DetailedException);                                \
+    }                                                                         \
+  while (0)
+
+/* ============================================================================
+ * Unified Error + Raise Macros (Eliminates Redundant Patterns)
+ * ============================================================================
+ */
+
+/**
+ * @brief SOCKET_RAISE_FMT - Format error with errno and raise exception in one
+ step
+ *
+ * @module_name: Module name for exception
+ * @exception: Exception to raise
+ * @fmt: Printf-style format string
+ * @...: Format arguments
+ *
+ * Combines SOCKET_ERROR_FMT + RAISE_MODULE_ERROR into single macro.
+ * @brief Thread-safe: Yes (uses thread-local buffers)
+ *
+ */
+#define SOCKET_RAISE_FMT(module_name, exception, fmt, ...)                    \
+  do                                                                          \
+    {                                                                         \
+      SOCKET_ERROR_FMT (fmt, ##__VA_ARGS__);                                  \
+      SOCKET_RAISE_MODULE_ERROR (module_name, exception);                     \
+    }                                                                         \
+  while (0)
+
+/**
+ * @brief SOCKET_RAISE_MSG - Format error message and raise exception in one
+ step
+ *
+ * @module_name: Module name for exception
+ * @exception: Exception to raise
+ * @fmt: Printf-style format string (without errno)
+ * @...: Format arguments
+ *
+ * Combines SOCKET_ERROR_MSG + RAISE_MODULE_ERROR into single macro.
+ * @brief Thread-safe: Yes (uses thread-local buffers)
+ *
+ */
+#define SOCKET_RAISE_MSG(module_name, exception, fmt, ...)                    \
+  do                                                                          \
+    {                                                                         \
+      SOCKET_ERROR_MSG (fmt, ##__VA_ARGS__);                                  \
+      SOCKET_RAISE_MODULE_ERROR (module_name, exception);                     \
+    }                                                                         \
+  while (0)
+
+/**
+ * Helper macros for common module patterns - use RAISE_MODULE_ERROR macro
+ * defined in each module that sets module_name appropriately.
+ *
+ * Example module setup:
+ *   SOCKET_DECLARE_MODULE_EXCEPTION(MyModule);
+ *   #define RAISE_MODULE_ERROR(e) SOCKET_RAISE_MODULE_ERROR(MyModule, e)
+ *   #define RAISE_FMT(e, fmt, ...) SOCKET_RAISE_FMT(MyModule, e, fmt,
+ * ##__VA_ARGS__) #define RAISE_MSG(e, fmt, ...) SOCKET_RAISE_MSG(MyModule, e,
+ * fmt, ##__VA_ARGS__)
+ */
+
+/* ============================================================================
+ * METRICS SUBSYSTEM (Legacy API)
  * ============================================================================
  */
 
@@ -545,470 +311,6 @@ SocketMetrics_snapshot_value (const SocketMetricsSnapshot *snapshot,
     return 0ULL;
   return snapshot->values[metric];
 }
-
-/* ============================================================================
- * EVENTS SUBSYSTEM
- * ============================================================================
- */
-
-/**
- * @brief SocketEventType - Event type enumeration
-
- */
-typedef enum SocketEventType
-{
-  SOCKET_EVENT_ACCEPTED = 0,
-  SOCKET_EVENT_CONNECTED,
-  SOCKET_EVENT_DNS_TIMEOUT,
-  SOCKET_EVENT_POLL_WAKEUP
-} SocketEventType;
-
-/**
- * @brief SocketEventRecord - Event data structure
-
- */
-typedef struct SocketEventRecord
-{
-  SocketEventType type;
-  const char *component;
-  union
-  {
-    struct
-    {
-      int fd;
-      const char *peer_addr;
-      int peer_port;
-      const char *local_addr;
-      int local_port;
-    } connection;
-    struct
-    {
-      const char *host;
-      int port;
-    } dns;
-    struct
-    {
-      int nfds;
-      int timeout_ms;
-    } poll;
-  } data;
-} SocketEventRecord;
-
-/**
- * @brief SocketEventCallback - Event handler callback type
-
- * @userdata: User-provided context
- * @event: Event record
- */
-typedef void (*SocketEventCallback) (void *userdata,
-                                     const SocketEventRecord *event);
-
-/**
- * @brief SocketEvent_register - Register an event handler
- * @ingroup foundation
- * @param callback Callback function to register
- * @param userdata User data passed to callback
- * @threadsafe Yes
- */
-void SocketEvent_register (SocketEventCallback callback, void *userdata);
-
-/**
- * @brief SocketEvent_unregister - Unregister an event handler
- * @ingroup foundation
- * @param callback Callback function to unregister
- * @param userdata User data that was passed to register
- * @threadsafe Yes
- */
-void SocketEvent_unregister (SocketEventCallback callback,
-                             const void *userdata);
-
-/**
- * @brief Emit connection accept event.
- * @ingroup foundation
- *
- * @param fd Client file descriptor.
- * @param peer_addr Peer IP address.
- * @param peer_port Peer port.
- * @param local_addr Local IP address.
- * @param local_port Local port.
- *
- * @threadsafe Yes
- */
-void SocketEvent_emit_accept (int fd, const char *peer_addr, int peer_port,
-                              const char *local_addr, int local_port);
-
-/**
- * @brief Emit outbound connection event.
- * @ingroup foundation
- *
- * @param fd Socket file descriptor.
- * @param peer_addr Peer IP address.
- * @param peer_port Peer port.
- * @param local_addr Local IP address.
- * @param local_port Local port.
- *
- * @threadsafe Yes
- */
-void SocketEvent_emit_connect (int fd, const char *peer_addr, int peer_port,
-                               const char *local_addr, int local_port);
-
-/**
- * @brief Emit DNS resolution timeout event.
- * @ingroup foundation
- *
- * @param host Hostname that timed out.
- * @param port Destination port.
- *
- * @threadsafe Yes
- */
-void SocketEvent_emit_dns_timeout (const char *host, int port);
-
-/**
- * @brief Emit poll wakeup event.
- * @ingroup foundation
- *
- * @param nfds Number of monitored file descriptors.
- * @param timeout_ms Poll timeout (-1 = infinite).
- *
- * @threadsafe Yes
- */
-void SocketEvent_emit_poll_wakeup (int nfds, int timeout_ms);
-
-/* ============================================================================
- * ERROR HANDLING SUBSYSTEM
- * ============================================================================
- */
-
-/**
- * @brief Normalized error codes mapping POSIX errno values.
- * @ingroup foundation
- */
-typedef enum SocketErrorCode
-{
-  SOCKET_ERROR_NONE = 0,
-  SOCKET_ERROR_EINVAL,
-  SOCKET_ERROR_EACCES,
-  SOCKET_ERROR_EADDRINUSE,
-  SOCKET_ERROR_EADDRNOTAVAIL,
-  SOCKET_ERROR_EAFNOSUPPORT,
-  SOCKET_ERROR_EAGAIN,
-  SOCKET_ERROR_EALREADY,
-  SOCKET_ERROR_EBADF,
-  SOCKET_ERROR_ECONNREFUSED,
-  SOCKET_ERROR_ECONNRESET,
-  SOCKET_ERROR_EFAULT,
-  SOCKET_ERROR_EHOSTUNREACH,
-  SOCKET_ERROR_EINPROGRESS,
-  SOCKET_ERROR_EINTR,
-  SOCKET_ERROR_EISCONN,
-  SOCKET_ERROR_EMFILE,
-  SOCKET_ERROR_ENETUNREACH,
-  SOCKET_ERROR_ENOBUFS,
-  SOCKET_ERROR_ENOMEM,
-  SOCKET_ERROR_ENOTCONN,
-  SOCKET_ERROR_ENOTSOCK,
-  SOCKET_ERROR_EOPNOTSUPP,
-  SOCKET_ERROR_EPIPE,
-  SOCKET_ERROR_EPROTONOSUPPORT,
-  SOCKET_ERROR_ETIMEDOUT,
-  SOCKET_ERROR_EWOULDBLOCK,
-  SOCKET_ERROR_UNKNOWN
-} SocketErrorCode;
-
-/* Thread-local error buffer for detailed messages */
-#ifdef _WIN32
-extern __declspec (thread) char socket_error_buf[SOCKET_ERROR_BUFSIZE];
-extern __declspec (thread) int socket_last_errno;
-#else
-extern __thread char socket_error_buf[SOCKET_ERROR_BUFSIZE];
-extern __thread int socket_last_errno;
-#endif
-
-/**
- * @brief SOCKET_ERROR_APPLY_TRUNCATION - Apply truncation marker if message
- was cut
-
- * @ret: Return value from snprintf
- *
- * Internal helper macro to eliminate duplication in error formatting.
- */
-#define SOCKET_ERROR_APPLY_TRUNCATION(ret)                                    \
-  do                                                                          \
-    {                                                                         \
-      if ((ret) >= (int)SOCKET_ERROR_BUFSIZE)                                 \
-        {                                                                     \
-          socket_error_buf[SOCKET_ERROR_BUFSIZE - 1] = '\0';                  \
-          if (SOCKET_ERROR_BUFSIZE >= SOCKET_ERROR_TRUNCATION_SIZE + 1)       \
-            {                                                                 \
-              memcpy (socket_error_buf + SOCKET_ERROR_BUFSIZE                 \
-                          - SOCKET_ERROR_TRUNCATION_SIZE,                     \
-                      SOCKET_ERROR_TRUNCATION_MARKER,                         \
-                      SOCKET_ERROR_TRUNCATION_SIZE - 1);                      \
-              socket_error_buf[SOCKET_ERROR_BUFSIZE - 1] = '\0';              \
-            }                                                                 \
-        }                                                                     \
-    }                                                                         \
-  while (0)
-
-/**
- * @brief SOCKET_ERROR_FMT - Format error message with errno information
-
- * Includes truncation protection for long messages.
- */
-#define SOCKET_ERROR_FMT(fmt, ...)                                            \
-  do                                                                          \
-    {                                                                         \
-      socket_last_errno = errno;                                              \
-      char tmp_buf[SOCKET_ERROR_BUFSIZE];                                     \
-      int _socket_error_ret = snprintf (                                      \
-          tmp_buf, sizeof (tmp_buf), fmt " (errno: %d - %s)", ##__VA_ARGS__,  \
-          socket_last_errno, Socket_safe_strerror (socket_last_errno));       \
-      memcpy (socket_error_buf, tmp_buf, SOCKET_ERROR_BUFSIZE);               \
-      socket_error_buf[SOCKET_ERROR_BUFSIZE - 1] = '\0';                      \
-      SOCKET_ERROR_APPLY_TRUNCATION (_socket_error_ret);                      \
-      (void)_socket_error_ret;                                                \
-      SocketLog_emit (SOCKET_LOG_ERROR, SOCKET_LOG_COMPONENT,                 \
-                      socket_error_buf);                                      \
-    }                                                                         \
-  while (0)
-
-/**
- * @brief SOCKET_ERROR_MSG - Format error message without errno
-
- * Includes truncation protection for long messages.
- */
-#define SOCKET_ERROR_MSG(fmt, ...)                                            \
-  do                                                                          \
-    {                                                                         \
-      socket_last_errno = errno;                                              \
-      char tmp_buf[SOCKET_ERROR_BUFSIZE];                                     \
-      int _socket_error_ret                                                   \
-          = snprintf (tmp_buf, sizeof (tmp_buf), fmt, ##__VA_ARGS__);         \
-      memcpy (socket_error_buf, tmp_buf, SOCKET_ERROR_BUFSIZE);               \
-      socket_error_buf[SOCKET_ERROR_BUFSIZE - 1] = '\0';                      \
-      SOCKET_ERROR_APPLY_TRUNCATION (_socket_error_ret);                      \
-      (void)_socket_error_ret;                                                \
-      SocketLog_emit (SOCKET_LOG_ERROR, SOCKET_LOG_COMPONENT,                 \
-                      socket_error_buf);                                      \
-    }                                                                         \
-  while (0)
-
-/**
- * @brief Get the last formatted error message.
- * @ingroup foundation
- *
- * @return Thread-local error string (never NULL).
- *
- * @threadsafe Yes
- */
-extern const char *Socket_GetLastError (void);
-
-/**
- * @brief Get the raw errno from the last error.
- * @ingroup foundation
- *
- * @return Last errno value.
- *
- * @threadsafe Yes
- */
-extern int Socket_geterrno (void);
-
-/**
- * @brief Convert last errno to normalized SocketErrorCode.
- * @ingroup foundation
- *
- * @return Mapped SocketErrorCode (SOCKET_ERROR_UNKNOWN if unmapped).
- *
- * @threadsafe Yes
- */
-extern SocketErrorCode Socket_geterrorcode (void);
-
-/**
- * @brief Thread-safe errno to string conversion.
- * @ingroup foundation
- *
- * @param errnum errno value to convert.
- * @return Descriptive error string.
- *
- * @threadsafe Yes
- */
-const char *Socket_safe_strerror (int errnum);
-
-/* Common error conditions with descriptive messages */
-#define SOCKET_ENOMEM "Out of memory"
-#define SOCKET_EINVAL "Invalid argument"
-#define SOCKET_ECONNREFUSED "Connection refused"
-#define SOCKET_ETIMEDOUT "Operation timed out"
-#define SOCKET_EADDRINUSE "Address already in use"
-#define SOCKET_ENETUNREACH "Network unreachable"
-#define SOCKET_EHOSTUNREACH "Host unreachable"
-#define SOCKET_EPIPE "Broken pipe"
-#define SOCKET_ECONNRESET "Connection reset by peer"
-
-/* ============================================================================
- * ERROR CATEGORIZATION
- * ============================================================================
- *
- * Provides error classification for determining retry eligibility and
- * appropriate error handling strategies.
- *
- * Categories:
- * - NETWORK: Transient network errors (usually retryable)
- * - PROTOCOL: Protocol/format errors (usually not retryable)
- * - APPLICATION: Application-level errors (context-dependent)
- * - TIMEOUT: Timeout errors (usually retryable with backoff)
- * - RESOURCE: Resource exhaustion (may be retryable after delay)
- * - UNKNOWN: Unclassified errors
- */
-
-/**
- * @brief High-level classification of error types.
- * @ingroup foundation
- */
-typedef enum SocketErrorCategory
-{
-  SOCKET_ERROR_CATEGORY_NETWORK
-  = 0, /**< Network-level: ECONNRESET, ECONNREFUSED, etc. */
-  SOCKET_ERROR_CATEGORY_PROTOCOL,    /**< Protocol-level: Parse errors, invalid
-                                        responses */
-  SOCKET_ERROR_CATEGORY_APPLICATION, /**< App-level: Auth failures, 4xx
-                                        responses */
-  SOCKET_ERROR_CATEGORY_TIMEOUT,     /**< Timeout errors: ETIMEDOUT, deadline
-                                        exceeded */
-  SOCKET_ERROR_CATEGORY_RESOURCE, /**< Resource exhaustion: OOM, fd limits */
-  SOCKET_ERROR_CATEGORY_UNKNOWN   /**< Unclassified errors */
-} SocketErrorCategory;
-
-/**
- * @brief Classify errno into SocketErrorCategory.
- * @ingroup foundation
- *
- * @param err errno value to classify.
- * @return Appropriate SocketErrorCategory.
- *
- * @threadsafe Yes
- */
-extern SocketErrorCategory SocketError_categorize_errno (int err);
-
-/**
- * @brief Get string name for error category.
- * @param category Error category.
- * @return Static string with category name.
- * @threadsafe Yes (returns static data)
- */
-extern const char *SocketError_category_name (SocketErrorCategory category);
-
-/**
- * @brief SocketError_is_retryable_errno - Check if errno indicates retryable
- error
-
- * @err: errno value to check
- *
- * Returns: 1 if error is typically retryable, 0 if fatal
- * @brief Thread-safe: Yes (pure function)
-
- *
- * Retryable errors include:
- * - Network transient: ECONNREFUSED, ECONNRESET, ENETUNREACH, EHOSTUNREACH
- * - Timeout: ETIMEDOUT
- * - Temporary resource: EAGAIN, EWOULDBLOCK, EINTR
- *
- * @brief Non-retryable errors include:
-
- * - Configuration: EACCES, EADDRINUSE, EADDRNOTAVAIL, EPERM
- * - Programming: EBADF, ENOTSOCK, EINVAL, EFAULT
- * - Permanent resource: ENOMEM, EMFILE, ENFILE
- */
-extern int SocketError_is_retryable_errno (int err);
-
-/* ============================================================================
- * Centralized Exception Infrastructure
- * ============================================================================
- */
-
-/**
- * @brief SOCKET_DECLARE_MODULE_EXCEPTION - Declare thread-local exception
-
- * @module_name: Module name (e.g., Socket, SocketBuf, SocketPoll)
- */
-#define SOCKET_DECLARE_MODULE_EXCEPTION(module_name)                          \
-  static __thread Except_T module_name##_DetailedException
-
-/**
- * @brief SOCKET_RAISE_MODULE_ERROR - Raise module-specific exception
-
- * @module_name: Module name
- * @exception: Exception to raise
- * @brief Thread-safe: Creates thread-local copy with detailed reason
-
- */
-#define SOCKET_RAISE_MODULE_ERROR(module_name, exception)                     \
-  do                                                                          \
-    {                                                                         \
-      module_name##_DetailedException = (exception);                          \
-      module_name##_DetailedException.reason = socket_error_buf;              \
-      RAISE (module_name##_DetailedException);                                \
-    }                                                                         \
-  while (0)
-
-/* ============================================================================
- * Unified Error + Raise Macros (Eliminates Redundant Patterns)
- * ============================================================================
- */
-
-/**
- * @brief SOCKET_RAISE_FMT - Format error with errno and raise exception in one
- step
-
- * @module_name: Module name for exception
- * @exception: Exception to raise
- * @fmt: Printf-style format string
- * @...: Format arguments
- *
- * Combines SOCKET_ERROR_FMT + RAISE_MODULE_ERROR into single macro.
- * @brief Thread-safe: Yes (uses thread-local buffers)
-
- */
-#define SOCKET_RAISE_FMT(module_name, exception, fmt, ...)                    \
-  do                                                                          \
-    {                                                                         \
-      SOCKET_ERROR_FMT (fmt, ##__VA_ARGS__);                                  \
-      SOCKET_RAISE_MODULE_ERROR (module_name, exception);                     \
-    }                                                                         \
-  while (0)
-
-/**
- * @brief SOCKET_RAISE_MSG - Format error message and raise exception in one
- step
-
- * @module_name: Module name for exception
- * @exception: Exception to raise
- * @fmt: Printf-style format string (without errno)
- * @...: Format arguments
- *
- * Combines SOCKET_ERROR_MSG + RAISE_MODULE_ERROR into single macro.
- * @brief Thread-safe: Yes (uses thread-local buffers)
-
- */
-#define SOCKET_RAISE_MSG(module_name, exception, fmt, ...)                    \
-  do                                                                          \
-    {                                                                         \
-      SOCKET_ERROR_MSG (fmt, ##__VA_ARGS__);                                  \
-      SOCKET_RAISE_MODULE_ERROR (module_name, exception);                     \
-    }                                                                         \
-  while (0)
-
-/**
- * Helper macros for common module patterns - use RAISE_MODULE_ERROR macro
- * defined in each module that sets module_name appropriately.
- *
- * Example module setup:
- *   SOCKET_DECLARE_MODULE_EXCEPTION(MyModule);
- *   #define RAISE_MODULE_ERROR(e) SOCKET_RAISE_MODULE_ERROR(MyModule, e)
- *   #define RAISE_FMT(e, fmt, ...) SOCKET_RAISE_FMT(MyModule, e, fmt,
- * ##__VA_ARGS__) #define RAISE_MSG(e, fmt, ...) SOCKET_RAISE_MSG(MyModule, e,
- * fmt, ##__VA_ARGS__)
- */
 
 /* ============================================================================
  * TIME UTILITIES (Consolidated monotonic clock functions)
